Check for null model in task edit onDetatch()

+++ astrid/src/com/todoroo/astrid/activity/TaskEditFragment.java
@@ -874,7 +874,8 @@ ViewPager.OnPageChangeListener, EditNoteActivity.UpdatesChangedListener {
                     R.anim.slide_right_in, R.anim.slide_right_out);
         }
 
-        if (title.getText().length() == 0 && isNewTask && model.isSaved()) {
+        if (title.getText().length() == 0 && isNewTask
+                && model != null && model.isSaved()) {
             taskService.delete(model);
         }
     }
